5547 - Add useExtraActions
